-FIX: Removed duplicate of MqttClientService and MqttOptionService, left the one in Convenience and added the RouterDevice-Part
